# backend/ecotrack-gateway/src/app.js
require('dotenv').config();
const express = require('express');
const cors = require('cors');
const helmet = require('helmet');
const axios = require('axios');
const fs = require('fs');
const path = require('path');
const yaml = require('yaml');
const swaggerUi = require('swagger-ui-express');

const app = express();
const PORT = process.env.GATEWAY_PORT || 3000;

// Service URLs
const SERVICES = {
  auth: process.env.AUTH_SERVICE_URL || 'http://localhost:3001',
  container: process.env.CONTAINER_SERVICE_URL || 'http://localhost:3002',
  tour: process.env.TOUR_SERVICE_URL || 'http://localhost:3003',
  signal: process.env.SIGNAL_SERVICE_URL || 'http://localhost:3004',
  iot: process.env.IOT_SERVICE_URL || 'http://localhost:3006',
  user: process.env.USER_SERVICE_URL || 'http://localhost:3005',
};

// ============ MIDDLEWARE ============

// Security headers
app.use(helmet());

// CORS
app.use(cors({
  origin: process.env.CORS_ORIGIN || '*',
  credentials: true,
  optionsSuccessStatus: 200,
}));

// Body parsing
app.use(express.json({ limit: '10mb' }));
app.use(express.urlencoded({ limit: '10mb', extended: true }));

// Request logging middleware
app.use((req, res, next) => {
  const start = Date.now();
  console.log(`[${new Date().toISOString()}] ${req.method} ${req.path}`);
  
  res.on('finish', () => {
    const duration = Date.now() - start;
    console.log(`[${new Date().toISOString()}] ${req.method} ${req.path} - ${res.statusCode} (${duration}ms)`);
  });
  
  next();
});

// ============ SWAGGER API DOCUMENTATION ============

try {
  const swaggerFile = fs.readFileSync(path.join(__dirname, '../../swagger.yaml'), 'utf8');
  const swaggerDoc = yaml.parse(swaggerFile);
  
  app.use('/api-docs', swaggerUi.serve);
  app.get('/api-docs', swaggerUi.setup(swaggerDoc, {
    swaggerOptions: {
      urls: [
        {
          url: '/swagger-spec',
          name: 'EcoTrack API'
        }
      ]
    }
  }));
  
  // Serve raw swagger spec
  app.get('/swagger-spec', (req, res) => {
    res.type('application/yaml').send(swaggerFile);
  });
  
  console.log('✓ Swagger UI available at http://localhost:3000/api-docs');
} catch (err) {
  console.warn('⚠ Swagger UI configuration failed:', err.message);
}

// ============ HEALTH CHECK ============

app.get('/health', async (req, res) => {
  try {
    const serviceHealth = {};
    
    // Check each service health
    for (const [name, url] of Object.entries(SERVICES)) {
      try {
        await axios.get(`${url}/health`, { timeout: 3000 });
        serviceHealth[name] = 'healthy';
      } catch (error) {
        serviceHealth[name] = 'unhealthy';
      }
    }

    const allHealthy = Object.values(serviceHealth).every(s => s === 'healthy');
    
    res.status(allHealthy ? 200 : 503).json({
      status: allHealthy ? 'healthy' : 'degraded',
      timestamp: new Date().toISOString(),
      services: serviceHealth,
    });
  } catch (error) {
    res.status(500).json({
      status: 'error',
      message: error.message,
    });
  }
});

// ============ PROXY FUNCTION ============

const proxyRequest = async (req, res, serviceUrl) => {
  try {
    const config = {
      method: req.method,
      url: `${serviceUrl}${req.path}`,
      params: req.query,
      headers: {
        ...req.headers,
        'host': new URL(serviceUrl).host,
      },
      data: req.method !== 'GET' && req.method !== 'HEAD' ? req.body : undefined,
      timeout: 30000,
    };

    // Forward the request to the service
    const response = await axios(config);
    
    // Return the response
    res.status(response.status).json(response.data);
  } catch (error) {
    if (error.response) {
      // Service returned an error
      res.status(error.response.status).json(error.response.data);
    } else if (error.code === 'ECONNREFUSED') {
      res.status(503).json({
        success: false,
        message: 'Service unavailable',
        error: 'Cannot connect to service',
      });
    } else if (error.code === 'ENOTFOUND') {
      res.status(503).json({
        success: false,
        message: 'Service not found',
        error: 'DNS resolution failed',
      });
    } else {
      res.status(500).json({
        success: false,
        message: 'Gateway error',
        error: error.message,
      });
    }
  }
};

// ============ ROUTES - AUTH SERVICE ============

app.post('/api/auth/register', (req, res) => proxyRequest(req, res, SERVICES.auth));
app.post('/api/auth/login', (req, res) => proxyRequest(req, res, SERVICES.auth));
app.post('/api/auth/logout', (req, res) => proxyRequest(req, res, SERVICES.auth));
app.post('/api/auth/verify', (req, res) => proxyRequest(req, res, SERVICES.auth));
app.post('/api/auth/refresh-token', (req, res) => proxyRequest(req, res, SERVICES.auth));
app.post('/api/auth/refresh', (req, res) => proxyRequest(req, res, SERVICES.auth));
app.get('/api/auth/me', (req, res) => proxyRequest(req, res, SERVICES.auth));

// ============ ROUTES - USER SERVICE ============

app.get('/api/users', (req, res) => proxyRequest(req, res, SERVICES.user));
app.get('/api/users/me', (req, res) => proxyRequest(req, res, SERVICES.user));
// Static sub-routes before /:id to avoid capture by the dynamic param
app.get('/api/users/:id/profile', (req, res) => proxyRequest(req, res, SERVICES.user));
app.put('/api/users/:id/role', (req, res) => proxyRequest(req, res, SERVICES.user));
app.get('/api/users/:id', (req, res) => proxyRequest(req, res, SERVICES.user));
app.put('/api/users/:id', (req, res) => proxyRequest(req, res, SERVICES.user));
app.delete('/api/users/:id', (req, res) => proxyRequest(req, res, SERVICES.user));

// POST /api/users → admin user creation via auth service register
app.post('/api/users', async (req, res) => {
  try {
    const response = await axios({
      method: 'POST',
      url: `${SERVICES.auth}/api/auth/register`,
      data: req.body,
      headers: {
        'Content-Type': 'application/json',
        ...(req.headers.authorization ? { Authorization: req.headers.authorization } : {}),
      },
      timeout: 30000,
    });
    res.status(response.status).json(response.data);
  } catch (error) {
    if (error.response) {
      res.status(error.response.status).json(error.response.data);
    } else if (error.code === 'ECONNREFUSED') {
      res.status(503).json({ success: false, message: 'Auth service unavailable' });
    } else {
      res.status(500).json({ success: false, message: 'Gateway error', error: error.message });
    }
  }
});

// ============ ROUTES - CONTAINER SERVICE ============

app.get('/api/zones', (req, res) => proxyRequest(req, res, SERVICES.container));
app.post('/api/zones', (req, res) => proxyRequest(req, res, SERVICES.container));
app.get('/api/zones/:id', (req, res) => proxyRequest(req, res, SERVICES.container));
app.put('/api/zones/:id', (req, res) => proxyRequest(req, res, SERVICES.container));
app.delete('/api/zones/:id', (req, res) => proxyRequest(req, res, SERVICES.container));

app.get('/api/conteneurs', (req, res) => proxyRequest(req, res, SERVICES.container));
app.post('/api/conteneurs', (req, res) => proxyRequest(req, res, SERVICES.container));
// Routes statiques avant /:id pour éviter la capture par le param dynamique
app.get('/api/conteneurs/needs-service', (req, res) => proxyRequest(req, res, SERVICES.container));
app.get('/api/conteneurs/nearby', (req, res) => proxyRequest(req, res, SERVICES.container));
app.get('/api/conteneurs/:id', (req, res) => proxyRequest(req, res, SERVICES.container));
app.put('/api/conteneurs/:id', (req, res) => proxyRequest(req, res, SERVICES.container));
app.delete('/api/conteneurs/:id', (req, res) => proxyRequest(req, res, SERVICES.container));

// ============ ROUTES - CAPTEURS (via container-service) ============

app.post('/api/capteurs', (req, res) => proxyRequest(req, res, SERVICES.container));
app.get('/api/capteurs', (req, res) => proxyRequest(req, res, SERVICES.container));
app.get('/api/capteurs/conteneur/:conteneurId', (req, res) => proxyRequest(req, res, SERVICES.container));
app.get('/api/capteurs/:id', (req, res) => proxyRequest(req, res, SERVICES.container));
app.put('/api/capteurs/:id', (req, res) => proxyRequest(req, res, SERVICES.container));
app.delete('/api/capteurs/:id', (req, res) => proxyRequest(req, res, SERVICES.container));
app.patch('/api/capteurs/:id/conteneur', (req, res) => proxyRequest(req, res, SERVICES.container));
app.patch('/api/capteurs/:id/batterie', (req, res) => proxyRequest(req, res, SERVICES.container));

// GET /api/iot/device/:capteur_id → résolu via container-service
app.get('/api/iot/device/:capteur_id', async (req, res) => {
  try {
    const response = await axios({
      method: 'GET',
      url: `${SERVICES.container}/api/capteurs/${req.params.capteur_id}`,
      headers: { ...req.headers, host: new URL(SERVICES.container).host },
      timeout: 10000,
    });
    res.status(response.status).json(response.data);
  } catch (error) {
    if (error.response) {
      res.status(error.response.status).json(error.response.data);
    } else {
      res.status(503).json({ success: false, message: 'Container service unavailable' });
    }
  }
});

app.get('/api/mesures', (req, res) => proxyRequest(req, res, SERVICES.container));
app.post('/api/mesures', (req, res) => proxyRequest(req, res, SERVICES.container));
app.get('/api/mesures/container/:containerId', (req, res) => proxyRequest(req, res, SERVICES.container));
app.get('/api/mesures/analytics/average-fill', (req, res) => proxyRequest(req, res, SERVICES.container));

// Container stats (optional - for detailed container page stats)
app.get('/api/container-stats/dashboard', (req, res) => proxyRequest(req, res, SERVICES.container));
app.get('/api/container-stats/breakdown/status', (req, res) => proxyRequest(req, res, SERVICES.container));
app.get('/api/container-stats/breakdown/type', (req, res) => proxyRequest(req, res, SERVICES.container));

// ============ ROUTES - TOUR SERVICE ============

app.get('/api/tournees', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.post('/api/tournees', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.get('/api/tournees/:id', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.put('/api/tournees/:id', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.delete('/api/tournees/:id', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.get('/api/tournees/agent/:agentId', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.post('/api/tournees/:id/agents', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.delete('/api/tournees/:id/agents/:agentId', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.patch('/api/tournees/:id/statut', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.get('/api/tournees/:id/stats', (req, res) => proxyRequest(req, res, SERVICES.tour));

// POST /api/tournees/:tourneeId/signalements → forward to signal-service
app.post('/api/tournees/:tourneeId/signalements', async (req, res) => {
  try {
    const response = await axios({
      method: 'POST',
      url: `${SERVICES.signal}/api/signalements`,
      data: { ...req.body, id_tournee: req.params.tourneeId },
      headers: {
        'Content-Type': 'application/json',
        ...(req.headers.authorization ? { Authorization: req.headers.authorization } : {}),
      },
      timeout: 30000,
    });
    res.status(response.status).json(response.data);
  } catch (error) {
    if (error.response) {
      res.status(error.response.status).json(error.response.data);
    } else if (error.code === 'ECONNREFUSED') {
      res.status(503).json({ success: false, message: 'Signal service unavailable' });
    } else {
      res.status(500).json({ success: false, message: 'Gateway error', error: error.message });
    }
  }
});

app.get('/api/collecteurs', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.post('/api/collecteurs', (req, res) => proxyRequest(req, res, SERVICES.tour));
// Routes statiques avant les routes dynamiques pour éviter que /:id ne capture "low-battery" ou "agent"
app.get('/api/collecteurs/low-battery', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.get('/api/collecteurs/agent/:agentId', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.get('/api/collecteurs/:id', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.put('/api/collecteurs/:id', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.delete('/api/collecteurs/:id', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.post('/api/collecteurs/:id/maintenance', (req, res) => proxyRequest(req, res, SERVICES.tour));

// Tour stats (optional - for detailed tour page stats)
app.get('/api/tour-stats/dashboard', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.get('/api/tour-stats/in-progress', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.get('/api/tour-stats/completed', (req, res) => proxyRequest(req, res, SERVICES.tour));
app.get('/api/tour-stats/breakdown/status', (req, res) => proxyRequest(req, res, SERVICES.tour));

// ============ ROUTES - SIGNAL SERVICE ============

app.get('/api/signalements', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.post('/api/signalements', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.get('/api/signalements/open', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.get('/api/signalements/:id', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.put('/api/signalements/:id', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.delete('/api/signalements/:id', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.get('/api/signalements/citoyen/:citoyenId', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.get('/api/signalements/container/:containerId', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.post('/api/signalements/:id/in-progress', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.post('/api/signalements/:id/close', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.post('/api/signalements/:id/reject', (req, res) => proxyRequest(req, res, SERVICES.signal));

// Signal stats (optional - for detailed signal page stats)
app.get('/api/signal-stats/dashboard', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.get('/api/signal-stats/open', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.get('/api/signal-stats/breakdown/status', (req, res) => proxyRequest(req, res, SERVICES.signal));
app.get('/api/signal-stats/breakdown/priority', (req, res) => proxyRequest(req, res, SERVICES.signal));

// ============ ROUTES - IOT SERVICE ============

app.post('/api/iot/measure', (req, res) => proxyRequest(req, res, SERVICES.iot));
app.post('/api/iot/device/register', (req, res) => proxyRequest(req, res, SERVICES.iot));
// GET /api/iot/device/:capteur_id → handled above (custom handler → container-service)
app.get('/api/iot/status', (req, res) => proxyRequest(req, res, SERVICES.iot));

// ============ DASHBOARD STATS ============

app.get('/api/dashboard/stats', async (req, res) => {
  try {
    const authHeader = req.headers.authorization ? { Authorization: req.headers.authorization } : {};
    const [containerStats, tourStats, signalStats] = await Promise.all([
      axios.get(`${SERVICES.container}/api/stats/dashboard`, { timeout: 5000, headers: authHeader }).catch(() => ({ data: { data: {} } })),
      axios.get(`${SERVICES.tour}/api/stats/dashboard`, { timeout: 5000, headers: authHeader }).catch(() => ({ data: { data: {} } })),
      axios.get(`${SERVICES.signal}/api/stats/dashboard`, { timeout: 5000, headers: authHeader }).catch(() => ({ data: { data: {} } })),
    ]);

    const aggregatedStats = {
      containers: containerStats.data?.data?.containers || 0,
      zones: containerStats.data?.data?.zones || 0,
      tours: tourStats.data?.data?.totalTours || 0,
      toursInProgress: tourStats.data?.data?.toursInProgress || 0,
      averageFillRate: containerStats.data?.data?.averageFillRate || 0,
      criticalContainers: containerStats.data?.data?.criticalContainers || 0,
      openSignals: signalStats.data?.data?.openSignals || 0,
      totalSignals: signalStats.data?.data?.totalSignals || 0,
      containerBreakdown: {
        status: containerStats.data?.data?.statusBreakdown || {},
        type: containerStats.data?.data?.typeBreakdown || {}
      },
      tourBreakdown: tourStats.data?.data?.statusBreakdown || {},
      signalBreakdown: {
        status: signalStats.data?.data?.statusBreakdown || {},
        priority: signalStats.data?.data?.priorityBreakdown || {}
      }
    };

    res.status(200).json({
      success: true,
      message: 'Dashboard statistics retrieved',
      timestamp: new Date().toISOString(),
      data: aggregatedStats
    });
  } catch (error) {
    console.error('Dashboard stats error:', error);
    res.status(500).json({
      success: false,
      message: 'Failed to retrieve dashboard statistics',
      error: error.message
    });
  }
});

// ============ ERROR HANDLING ============

// 404 handler
app.use((req, res) => {
  res.status(404).json({
    success: false,
    message: 'Endpoint not found',
    path: req.path,
    method: req.method,
  });
});

// Global error handler
app.use((err, req, res, next) => {
  console.error('Error:', err);
  
  res.status(500).json({
    success: false,
    message: 'Internal server error',
    error: process.env.NODE_ENV === 'development' ? err.message : 'Server error',
  });
});

// ============ START SERVER ============

app.listen(PORT, '0.0.0.0', () => {
  console.log(`
╔═══════════════════════════════════════╗
║   EcoTrack API Gateway                ║
║   Port: ${PORT}                       ║
║   Status: RUNNING                     ║
║   ABDOU AZIZ BA                       ║
║   MOUNIB                              ║
║   VALENTIN                            ║
╚═══════════════════════════════════════╝
  `);
  
  console.log('Services configured:');
  Object.entries(SERVICES).forEach(([name, url]) => {
    console.log(`  ✓ ${name}: ${url}`);
  });
});

module.exports = app;