# backend/swagger.yaml
openapi: 3.0.0
info:
  title: EcoTrack API
  description: |
    Plateforme Event-Driven Microservices pour gestion optimisée de déchets urbains et collecte intelligente.
    
    ## Architecture
    - **API Gateway**: Point d'entrée unique sur http://localhost:3000
    - **6 Microservices**: auth, container, tour, signal, user, iot
    - **Message Broker**: RabbitMQ pour communication événementielle
    - **Database**: PostgreSQL par service (Database per Service pattern)
  version: 1.0.0
  contact:
    name: EcoTrack Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development Server (Gateway)
  - url: http://localhost:3001
    description: Auth Service
  - url: http://localhost:3002
    description: Container Service
  - url: http://localhost:3003
    description: Tour Service
  - url: http://localhost:3004
    description: Signal Service
  - url: http://localhost:3005
    description: User Service
  - url: http://localhost:3006
    description: IoT Service

# ============================================
# SECURITY SCHEMES
# ============================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenu via POST /auth/login
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [super_admin, admin, agent, citoyen]
        created_at:
          type: string
          format: date-time
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create new user account (public endpoint)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
              required:
                - email
                - password
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized
  /auth/verify:
    post:
      tags:
        - Authentication
      summary: Verify JWT token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid token
  /auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: New token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Profil de l'utilisateur connecté
      description: Retourne les informations décodées du JWT (id, email, role, nom, prenom).
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Informations de l'utilisateur connecté
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Token manquant ou invalide

  # ========== CONTAINER ENDPOINTS ==========
  /conteneurs:
    post:
      tags:
        - Container
      summary: Créer un conteneur
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code_conteneur
                - type_conteneur
                - capacite
                - latitude
                - longitude
                - date_installation
                - id_zone
              properties:
                code_conteneur:
                  type: string
                  example: CONT-2024-001
                type_conteneur:
                  type: string
                  enum: [standard, selective, organic, hazardous]
                capacite:
                  type: number
                  description: Capacité en litres
                latitude:
                  type: number
                longitude:
                  type: number
                date_installation:
                  type: string
                  format: date-time
                id_zone:
                  type: string
                  format: uuid
                notes:
                  type: string
      responses:
        '201':
          description: Conteneur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conteneur'
        '400':
          description: Données invalides
        '401':
          description: Non authentifié

    get:
      tags:
        - Container
      summary: Liste tous les conteneurs
      security:
        - BearerAuth: []
      parameters:
        - name: statut
          in: query
          schema:
            type: string
            enum: [actif, maintenance, retire]
        - name: id_zone
          in: query
          schema:
            type: string
            format: uuid
        - name: type_conteneur
          in: query
          schema:
            type: string
            enum: [standard, selective, organic, hazardous]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Liste des conteneurs
          content:
            application/json:
              schema:
                type: object
                properties:
                  conteneurs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conteneur'
                  total:
                    type: integer
  /conteneurs/nearby:
    get:
      tags:
        - Container
      summary: Conteneurs proches d'un point GPS
      description: Retourne les conteneurs actifs dans un rayon donné (Haversine). Max 50 résultats.
      security:
        - BearerAuth: []
      parameters:
        - name: lat
          in: query
          required: true
          description: Latitude du point de référence
          schema:
            type: number
            example: 48.8566
        - name: lng
          in: query
          required: true
          description: Longitude du point de référence
          schema:
            type: number
            example: 2.3522
        - name: radius
          in: query
          description: Rayon de recherche en kilomètres (défaut 5)
          schema:
            type: number
            default: 5
            example: 2
      responses:
        '200':
          description: Conteneurs trouvés dans le rayon
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  conteneurs:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Conteneur'
                        - type: object
                          properties:
                            distance_km:
                              type: number
                              description: Distance en km depuis le point de référence
        '400':
          description: lat et lng obligatoires ou coordonnées invalides
        '401':
          description: Non authentifié
  /conteneurs/needs-service:
    get:
      tags:
        - Container
      summary: Conteneurs nécessitant une intervention
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Conteneurs à traiter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conteneur'
  /conteneurs/{conteneurId}:
    get:
      tags:
        - Container
      summary: Détails d'un conteneur
      security:
        - BearerAuth: []
      parameters:
        - name: conteneurId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Détails du conteneur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conteneur'
        '404':
          description: Conteneur introuvable

    put:
      tags:
        - Container
      summary: Mettre à jour un conteneur
      security:
        - BearerAuth: []
      parameters:
        - name: conteneurId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                statut:
                  type: string
                  enum: [actif, maintenance, retire]
                capacite:
                  type: number
                notes:
                  type: string
      responses:
        '200':
          description: Conteneur mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conteneur'

    delete:
      tags:
        - Container
      summary: Supprimer un conteneur
      security:
        - BearerAuth: []
      parameters:
        - name: conteneurId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Conteneur supprimé

  # ========== MEASUREMENT ENDPOINTS ==========
  /mesures:
    post:
      tags:
        - Measurement
      summary: Enregistrer une mesure de remplissage
      description: Créer une nouvelle mesure pour un conteneur. Utilisé par le service IoT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - taux_remplissage
                - id_conteneur
              properties:
                id_conteneur:
                  type: string
                  format: uuid
                taux_remplissage:
                  type: number
                  minimum: 0
                  maximum: 100
                temperature:
                  type: number
                batterie:
                  type: integer
                  minimum: 0
                  maximum: 100
                signal_force:
                  type: integer
      responses:
        '201':
          description: Mesure enregistrée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mesure'
...
  /zones:
    post:
      tags:
        - Zone
      summary: Créer une zone
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nom
                - code_zone
              properties:
                nom:
                  type: string
                code_zone:
                  type: string
                geometrie:
                  type: string
                  description: GeoJSON (optionnel)
                population_estimee:
                  type: integer
      responses:
        '201':
          description: Zone créée

    get:
      tags:
        - Zone
      summary: Liste toutes les zones
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des zones
  /zones/{zoneId}:
    get:
      tags:
        - Zone
      summary: Détails d'une zone
      security:
        - BearerAuth: []
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Détails de la zone
        '404':
          description: Zone introuvable

    put:
      tags:
        - Zone
      summary: Mettre à jour une zone
      security:
        - BearerAuth: []
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Zone mise à jour

    delete:
      tags:
        - Zone
      summary: Supprimer une zone
      security:
        - BearerAuth: []
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Zone supprimée

  # ========== CONTAINER STATS ==========
  /container-stats/dashboard:
    get:
      tags:
        - Container Stats
      summary: Statistiques globales des conteneurs
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Statistiques dashboard conteneurs
          content:
            application/json:
              schema:
                type: object
                properties:
                  containers:
                    type: integer
                  zones:
                    type: integer
                  averageFillRate:
                    type: number
                  criticalContainers:
                    type: integer
  /container-stats/breakdown/status:
    get:
      tags:
        - Container Stats
      summary: Répartition des conteneurs par statut
      security:
        - BearerAuth: []
      responses:
        '200':
          description: "Répartition actif / maintenance / retire"
  /container-stats/breakdown/type:
    get:
      tags:
        - Container Stats
      summary: Répartition des conteneurs par type
      security:
        - BearerAuth: []
      responses:
        '200':
          description: "Répartition standard / selective / organic / hazardous"

  # ========== TOUR ENDPOINTS ==========
...
  /tournees/agent/{agentId}:
    get:
      tags:
        - Tour
      summary: Tournées d'un agent
      security:
        - BearerAuth: []
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tournées de l'agent
...
  /tournees/{id}/stats:
    get:
      tags:
        - Tour
      summary: Statistiques d'une tournée
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Statistiques de la tournée
...